+++ Pod/Classes/TBSMTransition.m
@@ -12,17 +12,17 @@
 @implementation TBSMTransition
 
 + (TBSMTransition *)transitionWithSourceState:(TBSMState *)sourceState
-                             targetState:(TBSMState *)targetState
-                                         type:(TBSMTransitionType)type
+                                  targetState:(TBSMState *)targetState
+                                         kind:(TBSMTransitionKind)kind
                                        action:(TBSMActionBlock)action
                                         guard:(TBSMGuardBlock)guard
 {
-    return [[TBSMTransition alloc] initWithSourceState:sourceState targetState:targetState type:type action:action guard:guard];
+    return [[TBSMTransition alloc] initWithSourceState:sourceState targetState:targetState kind:kind action:action guard:guard];
 }
 
 - (instancetype)initWithSourceState:(TBSMState *)sourceState
-                   targetState:(TBSMState *)targetState
-                               type:(TBSMTransitionType)type
+                        targetState:(TBSMState *)targetState
+                               kind:(TBSMTransitionKind)kind
                              action:(TBSMActionBlock)action
                               guard:(TBSMGuardBlock)guard
 {
@@ -30,7 +30,7 @@
     if (self) {
         _sourceState = sourceState;
         _targetState = targetState;
-        _type = type;
+        _kind = kind;
         _action = action;
         _guard = guard;
     }